Expr::New - Some(empty args) is equivalent to None
Based on the PartialEq implementation for Expr, a New variant with an
empty list of args is equivalent to a New variant with no args list
(`None`). Since initializing an empty vec does no allocation, we can
just use the equivalent representation and remove the `Option`
altogether.

Further, we can remove the custom implementaion of PartialEq since this
is the only thing that isn't equivalent.